pyproject.toml
[tool.poetry]
name = "html-mutation"
version = "0.1.0"
description = ""
authors = ["kabinja <renaud.rwemalika@gmail.com>"]
license = "Apache License 2.0"

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.2.0"
Pillow = "^9.1.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
